+++ isort/settings.py
@@ -17,7 +17,7 @@ from warnings import warn
 from . import stdlibs
 from ._future import dataclass, field
 from ._vendored import toml
-from .exceptions import InvalidSettingsPath, ProfileDoesNotExist
+from .exceptions import FormattingPluginDoesNotExist, InvalidSettingsPath, ProfileDoesNotExist
 from .profiles import profiles
 from .sections import DEFAULT as SECTION_DEFAULTS
 from .sections import FIRSTPARTY, FUTURE, LOCALFOLDER, STDLIB, THIRDPARTY
@@ -173,6 +173,8 @@ class _Config:
     remove_redundant_aliases: bool = False
     float_to_top: bool = False
     filter_files: bool = False
+    formatter: str = ""
+    formatting_function: Optional[Callable[[str, str, object], str]] = None
 
     def __post_init__(self):
         py_version = self.py_version
@@ -347,6 +349,16 @@ class Config(_Config):
             combined_config["src_paths"] = frozenset(
                 path_root / path for path in combined_config.get("src_paths", ())
             )
+
+        if "formatter" in combined_config:
+            import pkg_resources
+
+            for plugin in pkg_resources.iter_entry_points("isort.formatters"):
+                if plugin.name == combined_config["formatter"]:
+                    combined_config["formatting_function"] = plugin.load()
+                    break
+            else:
+                raise FormattingPluginDoesNotExist(combined_config["formatter"])
 
         # Remove any config values that are used for creating config object but
         # aren't defined in dataclass