+++ CHANGELOG.md
@@ -13,6 +13,7 @@ NOTE: isort follows the [semver](https://semver.org/) versioning standard.
   - Implemented #1214: Added support for git_hook lazy option (Thanks @sztamas!)
   - Implemented #941: Added an additional `multi_line_output` mode for more compact formatting (Thanks @sztamas!)
   - Implemented #1020: Option for LOCALFOLDER.
+  - Implemented #1353: Added support for output formatting plugins.
   - `# isort: split` can now be used at the end of an import line.
   - Fixed #1339: Extra indent is not preserved when isort:skip is used in nested imports.
   - Fixed #1348: `--diff` works incorrectly with files that have CRLF line endings.